requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
django-cors-headers==4.3.1
whitenoise==6.6.0
gunicorn==21.2.0

